+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Text.extend_style method.
+- Added Span.extend method.
 
 ## [13.4.2] - 2023-06-12